Protect critical sections better with mutexes
As this integration has grown, we need to be careful about what is and
is not protected. All calls exposed through the NIF that interact with a
statement or connection, need to acquire a mutex lock before operating
on the database. Nested helper function calls should refrain from
locking anything, otherwise we'll wind up with a deadlock.