docs: update VSCode terminal configuration for Windows users (#339)
Currently, Visual Studio Code's `terminal.integrated.shell.*` and
`terminal.integrated.shellArgs.*` are deprecated.

So I tried `terminal.integrated.profiles.windows` to replace source
configurations, and I found that it can still run normally if it is not
installed in the default directory (the specific situation has not been
studied in depth).